+++ dart/lib/src/diagnostic_logger.dart
@@ -7,9 +7,14 @@ class DiagnosticLogger {
 
   DiagnosticLogger(this._logger, this._options);
 
-  void log(SentryLevel level, String message) {
+  void log(
+    SentryLevel level,
+    String message, {
+    Object? error,
+    StackTrace? stackTrace,
+  }) {
     if (_isEnabled(level)) {
-      _logger(level, message);
+      _logger(level, message, error: error, stackTrace: stackTrace);
     }
   }